Hard-coded weather_clock.py into kill script and shutdown switch. Added check in kill script to make sure a PID was found

+++ kill_weather_clock.sh
@@ -1,10 +1,17 @@
 #!/bin/bash
 
-# name of weather clock python program
-weather_clock_py="$1"
-
 # get pid of weather clock process
-weather_clock_pid=$(ps -ef | grep "$weather_clock_py" | grep "root" | awk '{print $2}')
+weather_clock_pid=$(ps -ef | grep "weather_clock.py" | \
+                    grep python | grep "root" | awk '{print $2}')
+
+# check to make sure it got a PID
+if [ -z "$weather_clock_pid" ]; then
+    echo "The PID for weather_clock.py could not be found"
+    echo "exiting."
+    exit 1
+fi
 
 # kill the process
+echo "Killing weather_clock.py..."
 sudo kill --signal SIGINT $weather_clock_pid
+echo "  --> Killed."
